+++ test/sandbox/test/base_test.rb
@@ -88,7 +88,6 @@ class ViewComponent::Base::UnitTest < Minitest::Test
   end
 
   def test_does_not_render_additional_newline_with_render_in
-    skip unless Rails::VERSION::MAJOR >= 7
     without_template_annotations do
       ActionView::Template::Handlers::ERB.strip_trailing_newlines = true
       rendered_output = Array.new(2) {
@@ -97,7 +96,7 @@ class ViewComponent::Base::UnitTest < Minitest::Test
       assert_includes rendered_output, "<span>Hello, world!</span><span>Hello, world!</span>"
     end
   ensure
-    ActionView::Template::Handlers::ERB.strip_trailing_newlines = false if Rails::VERSION::MAJOR >= 7
+    ActionView::Template::Handlers::ERB.strip_trailing_newlines = false
   end
 
   def test_evaled_component