+++ test/sandbox/test/integration_test.rb
@@ -14,15 +14,13 @@ class IntegrationTest < ActionDispatch::IntegrationTest
     assert_select("div", "Foo\n  bar")
   end
 
-  if Rails.version.to_f >= 6.1
-    def test_rendering_component_with_template_annotations_enabled
-      get "/"
-      assert_response :success
+  def test_rendering_component_with_template_annotations_enabled
+    get "/"
+    assert_response :success
 
-      assert_includes response.body, "BEGIN app/components/erb_component.html.erb"
+    assert_includes response.body, "BEGIN app/components/erb_component.html.erb"
 
-      assert_select("div", "Foo\n  bar")
-    end
+    assert_select("div", "Foo\n  bar")
   end
 
   def test_rendering_component_in_a_controller
@@ -493,42 +491,6 @@ class IntegrationTest < ActionDispatch::IntegrationTest
     assert_response :success
   end
 
-  if Rails.version.to_f >= 6.1
-    def test_rendering_component_using_the_render_component_helper_raises_an_error
-      error =
-        assert_raises ActionView::Template::Error do
-          get "/render_component"
-        end
-
-      matcher = (RUBY_VERSION >= "3.4") ? /undefined method 'render_component'/ : /undefined method `render_component' for/
-      assert_match(matcher, error.message)
-    end
-  end
-
-  if Rails.version.to_f < 6.1
-    def test_rendering_component_using_render_component
-      get "/render_component"
-      assert_includes response.body, "bar"
-    end
-
-    def test_rendering_component_in_a_controller_using_render_component
-      get "/controller_inline_render_component"
-      assert_includes response.body, "bar"
-    end
-
-    def test_rendering_component_in_a_controller_using_render_component_to_string
-      get "/controller_to_string_render_component"
-      assert_includes response.body, "bar"
-    end
-
-    def test_rendering_component_in_preview_using_render_component_and_monkey_patch_disabled
-      with_render_monkey_patch_config(false) do
-        get "/rails/view_components/monkey_patch_disabled_component/default"
-        assert_includes response.body, "<div>hello,world!</div>"
-      end
-    end
-  end
-
   def test_renders_the_inline_component_preview_examples_with_default_behaviour_and_with_their_own_templates
     get "/rails/view_components/inline_component/default"
     assert_select "input" do
@@ -549,25 +511,23 @@ class IntegrationTest < ActionDispatch::IntegrationTest
   end
 
   def test_does_not_render_additional_newline
-    skip unless Rails::VERSION::MAJOR >= 7
     without_template_annotations do
       ActionView::Template::Handlers::ERB.strip_trailing_newlines = true
       get "/rails/view_components/display_inline_component/with_newline"
       assert_includes response.body, "<span>Hello, world!</span><span>Hello, world!</span>"
     end
   ensure
-    ActionView::Template::Handlers::ERB.strip_trailing_newlines = false if Rails::VERSION::MAJOR >= 7
+    ActionView::Template::Handlers::ERB.strip_trailing_newlines = false
   end
 
   def test_does_not_render_additional_newline_with_render_in
-    skip unless Rails::VERSION::MAJOR >= 7
     without_template_annotations do
       ActionView::Template::Handlers::ERB.strip_trailing_newlines = true
       get "/rails/view_components/display_inline_component/with_newline_render_in"
       assert_includes response.body, "<span>Hello, world!</span><span>Hello, world!</span>"
     end
   ensure
-    ActionView::Template::Handlers::ERB.strip_trailing_newlines = false if Rails::VERSION::MAJOR >= 7
+    ActionView::Template::Handlers::ERB.strip_trailing_newlines = false
   end
 
   # This test documents a bug that reports an incompatibility with the turbo-rails gem's `turbo_stream` helper.
@@ -617,8 +577,6 @@ class IntegrationTest < ActionDispatch::IntegrationTest
   end
 
   def test_renders_an_inline_component_preview_using_a_haml_template
-    skip if Rails::VERSION::STRING < "6.1"
-
     get "/rails/view_components/inline_component/with_haml"
     assert_select "h1", "Some HAML here"
     assert_select "input[name=?]", "name"
@@ -631,8 +589,6 @@ class IntegrationTest < ActionDispatch::IntegrationTest
   end
 
   def test_renders_a_mix_of_haml_and_erb
-    skip if Rails::VERSION::STRING < "6.1"
-
     get "/nested_haml"
     assert_response :success
     assert_select ".foo > .bar > .baz > .quux > .haml-div"
@@ -647,8 +603,6 @@ class IntegrationTest < ActionDispatch::IntegrationTest
   end
 
   def test_renders_a_preview_template_using_haml_params_from_url_custom_template_and_locals
-    skip if Rails::VERSION::STRING < "6.1"
-
     get "/rails/view_components/inline_component/with_several_options?form_title=Title from params"
 
     assert_select "form" do